+++ src/strojure/ring_undertow/handler.clj
@@ -11,10 +11,13 @@
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
 (defn ring-sync
-  "Returns HttpHandler for **synchronous** ring handler function.
+  "Returns HttpHandler for **synchronous** [ring handler function][1].
 
   The function `handler-fn` takes one argument, a map representing a HTTP
-  request, and return a map representing the HTTP response."
+  request, and return a map representing the HTTP response.
+
+  [1]: https://github.com/ring-clojure/ring/wiki/Concepts#handlers
+  "
   [handler-fn]
   (handler/dispatch
     (reify HttpHandler
@@ -26,10 +29,13 @@
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 
 (defn ring-async
-  "Returns HttpHandler for **asynchronous** ring handler function.
+  "Returns HttpHandler for **asynchronous** [ring handler function][1].
 
   The function `handler-fn` takes three arguments: the request map, a response
-  callback and an exception callback."
+  callback and an exception callback.
+
+  [1]: https://github.com/ring-clojure/ring/wiki/Concepts#handlers
+  "
   [handler-fn]
   (reify HttpHandler
     (handleRequest [_ exchange]
@@ -37,5 +43,31 @@
         (handler-fn (request/build-request exchange)
                     (fn handle-async [response] (response/handle-response response exchange))
                     (partial exchange/async-throw exchange))))))
+
+;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+
+(defn ring
+  "Returns HttpHandler for [ring handler function][1].
+
+  - `:fn`    The function to return `HttpHandler` for.
+  - `:async` Boolean flag if handler function is synchronous or asynchronous.
+             See also documentation for [[ring-sync]] and [[ring-async]].
+
+  Can be used declaratively in server-configuration:
+
+      (server/start {:handler {:type ring, :fn my-handler-fn, :async true}})
+
+  NOTE: The `strojure.ring-undertow.handler` namespace need to be imported to
+        make declarative configuration available.
+
+  [1]: https://github.com/ring-clojure/ring/wiki/Concepts#handlers
+  "
+  [{handler :fn async :async}]
+  (assert (fn? handler) (str "Requires function in :handler key: " (pr-str handler)))
+  (if async
+    (ring-async handler)
+    (ring-sync handler)))
+
+(handler/define-type ring {:alias ::ring :as-handler ring})
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,